+++ worker/include/RTC/DirectTransport.hpp
@@ -23,6 +23,8 @@ namespace RTC
 		void SendRtcpPacket(RTC::RTCP::Packet* packet) override;
 		void SendRtcpCompoundPacket(RTC::RTCP::CompoundPacket* packet) override;
 		void SendSctpData(const uint8_t* data, size_t len) override;
+		void RecvStreamClosed(uint32_t ssrc) override;
+		void SendStreamClosed(uint32_t ssrc) override;
 	};
 } // namespace RTC