+++ worker/include/RTC/Transport.hpp
@@ -156,6 +156,8 @@ namespace RTC
 		virtual void SendRtcpPacket(RTC::RTCP::Packet* packet)                 = 0;
 		virtual void SendRtcpCompoundPacket(RTC::RTCP::CompoundPacket* packet) = 0;
 		virtual void SendSctpData(const uint8_t* data, size_t len)             = 0;
+		virtual void RecvStreamClosed(uint32_t ssrc)                           = 0;
+		virtual void SendStreamClosed(uint32_t ssrc)                           = 0;
 		void DistributeAvailableOutgoingBitrate();
 		void ComputeOutgoingDesiredBitrate(bool forceBitrate = false);
 		void EmitTraceEventProbationType(RTC::RtpPacket* packet) const;